+++ HFMonitor/src/sdr/perseus/fx2_control.cpp
@@ -201,7 +201,7 @@ namespace Perseus {
     }
 
   protected:
-    virtual eeprom_data read_eeprom_lowlevel(boost::uint16_t addr, size_t length) {
+    virtual eeprom_data read_eeprom_lowlevel(boost::uint16_t addr, uint8_t length) {
       int transferred(0);
       const EEPROM::cmd cmd = {CMD::eeprom_read, addr, uint8_t(length)};
       ASSERT_THROW(_usb_control->submit_bulk(EndPoint::cmd,
